added auth and changed crud (no join operations)

+++ backend/controllers/tasks.js
@@ -10,11 +10,12 @@ const getAllTasks = async (req, res) => {
     const status = req.query.status;
     const priority = req.query.priority;
 
-    let text = "select * from tasks";
-    let values = [];
+    const userId = req.user.userId;
+    let text = "select * from tasks where userid = $1";
+    let values = [userId];
 
     if (status) {
-        text += ` where status = $1`;
+        text += ` where status = $2`;
         values.push(status);
     }
     if (priority) {
@@ -27,46 +28,69 @@ const getAllTasks = async (req, res) => {
     values.push(skip);
 
     let result = await client.query(text, values);
-    res.json({ nBHits: result.rowCount, data: result.rows });
+    res.status(200).json({ nBHits: result.rowCount, data: result.rows });
 };
 
 const getTask = async (req, res) => {
     const id = req.params.id;
+
     const query = {
         text: "select * from tasks where id = $1",
         values: [id],
     };
-    let result = await client.query(query);
+    const result = await client.query(query);
+    const row = result.rows[0];
 
-    res.json(result.rows[0]);
+    const userId = req.user.userId;
+    if (row.userid != userId) {
+        return res.status(401).json({ msg: "not authorized." });
+    }
+
+    res.status(200).json(row);
 };
 
 const createTask = async (req, res) => {
     const { title, desc, status, priority } = req.body;
+    const userId = req.user.userId;
     const query = {
-        text: "insert into tasks(title, description, status, priority) values($1, $2, $3, $4)",
-        values: [title, desc, status, priority],
+        text: "insert into tasks(title, description, status, priority, userId) values($1, $2, $3, $4, $5)",
+        values: [title, desc, status, priority, userId],
     };
     await client.query(query);
 
-    res.send("task created.");
+    res.status(200).json({ msg: "task created." });
 };
 
 const updateTask = async (req, res) => {
     const { title, desc, status, priority } = req.body;
     const id = req.params.id;
 
+    let result = await client.query("select * from tasks where id = " + id);
+    let row = result.rows[0];
+    const userId = req.user.userId;
+    if (row.userid != userId) {
+        return res.status(401).json({ msg: "not authorized." });
+    }
+
     const query = {
         text: "update tasks set title = $1, description = $2, status = $3, priority = $4 where id = $5",
         values: [title, desc, status, priority, id],
     };
-    const result = await client.query(query);
+    result = await client.query(query);
+    row = result.rows[0];
 
-    res.json({ msg: "task updated." });
+    res.status(200).json({ msg: "task updated." });
 };
 
 const deleteTask = async (req, res) => {
     const id = req.params.id;
+
+    let result = await client.query("select * from tasks where id = " + id);
+    let row = result.rows[0];
+    const userId = req.user.userId;
+    if (row.userid != userId) {
+        return res.status(401).json({ msg: "not authorized." });
+    }
 
     const query = {
         text: "delete from tasks where id = $1",
